Match rate limit value extraction to API responses
The API now serves these headers with a lower-case "l" on "limit", with
the result that we're not able to track whether we're close to hitting
the rate limit.

Rather than ask Dacast to update the API and potentially break things,
I've decided to fork and update their gem to match the API's current
behaviour.